+++ src/BruteForce.h
@@ -2,5 +2,5 @@
 
 class BruteForce {
   public:
-    static uint16_t Knapsack(uint8_t n, uint16_t M, Item *items, TaskType type, int16_t B);
+    static uint32_t Knapsack(uint8_t n, uint16_t M, Item *items, TaskType type, int32_t B);
 };
